examples/requirements.txt
# Core dependencies
automcp>=0.1.0
numpy>=1.24.0
pydantic>=2.0.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=4.1.0

# Optional dependencies
pandas>=2.0.0  # For advanced data manipulation
scipy>=1.12.0  # For additional statistical functions
matplotlib>=3.8.0  # For plotting capabilities
